Fix old file sent to wrong destination.

+++ src/collision/WorldModel.cpp
@@ -145,15 +145,15 @@ namespace VMAP
     bool WmoLiquid::GetLiquidHeight(const Vector3 &pos, float &liqHeight) const
     {
         float tx_f = (pos.x - iCorner.x)/LIQUID_TILE_SIZE;
-        uint32 tx = uint32(tx_f);
+        int32 tx = int32(tx_f);
 
-        if (tx_f < 0.0f || tx >= iTilesX)
+        if (tx_f < 0.0f || tx >= int32(iTilesX))
             return false; 
 
         float ty_f = (pos.y - iCorner.y)/LIQUID_TILE_SIZE;
-        uint32 ty = uint32(ty_f);
+        int32 ty = int32(ty_f);
 
-        if (ty_f < 0.0f || ty >= iTilesY)
+        if (ty_f < 0.0f || ty >= int32(iTilesY))
             return false; 
 
         // check if tile shall be used for liquid level
